ref: handle some edge cases

+++ src/services/organization.ts
@@ -1,4 +1,5 @@
 import { OrganizationDTO, QueryDTO } from '@app/data/models';
+import { orderProductRepo } from '@app/data/repositories/orderProduct.repo';
 import { orgRepo } from '@app/data/repositories/organization.repo';
 
 class OrganizationService {
@@ -15,6 +16,10 @@ class OrganizationService {
   }
 
   async deleteOrganization(id: string) {
+    const orders: any = await orderProductRepo.all({ where: { organizationId: id }})
+    if (orders?.length != 0) {
+      throw new Error("you can't delete an organization with orders")
+    }
     return await orgRepo.deleteByID(id);
   }
